# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: SPC 経費精算システム API
  description: |
    SPC経費精算システムのREST API仕様書
    
    ## 認証
    すべてのAPIリクエストには認証トークンが必要です。
    JWTトークンを `Authorization: Bearer {token}` ヘッダーで送信してください。
    
    ## エラーレスポンス
    エラー時は以下の形式でレスポンスを返します：
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "エラーメッセージ",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: SPC開発チーム
servers:
  - url: https://api.spc-expense.example.com/v1
    description: 本番環境
  - url: https://api-staging.spc-expense.example.com/v1
    description: ステージング環境
  - url: http://localhost:3000/api/v1
    description: ローカル開発環境
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - auth
      summary: トークンリフレッシュ
      description: リフレッシュトークンを使用して新しいアクセストークンを取得
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: リフレッシュトークン
      responses:
        '200':
          description: トークン更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: トークン無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT認証トークン
  schemas:
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: アクセストークン
        refreshToken:
          type: string
          description: リフレッシュトークン
        expiresIn:
          type: integer
          description: トークンの有効期限（秒）
        member:
          $ref: '#/components/schemas/Member'

    # 会員
    Member:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: string
          description: 会員ID（SPC会員DB連携）
        name:
          type: string
          description: 氏名
        email:
          type: string
          format: email
        departmentId:
          type: integer
        department:
          $ref: '#/components/schemas/Department'
        role:
          type: string
          enum: [member, admin, manager]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "バリデーションエラーが発生しました"
            details:
              type: object
              additionalProperties: true